\name{interest-package}
\alias{interest-package}
\alias{interest}
\docType{package}
\title{
\packageTitle{interest}
}
\description{
\packageDescription{interest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{interest}
\packageIndices{interest}
To do.....
}
\author{
\packageAuthor{interest}

Maintainer: \packageMaintainer{interest}
}
\references{
F. Mercurio: Interest Rates
}

\keyword{ package }
\seealso{}
\examples{
to follow
}
